moved ingest_wrapper to queue_service.py
