+++ ai_ta_backend/utils/OpenaiEmbeddings.py
